+++ mockzilla-management-ui/mockzilla-management-ui-common/src/commonMain/kotlin/com/apadmi/mockzilla/ui/ui/common/components/editor/EditorTextField.kt
@@ -50,7 +50,6 @@ import androidx.compose.ui.input.key.onKeyEvent
 import androidx.compose.ui.input.key.type
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextLayoutResult
 import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.TextStyle
@@ -66,9 +65,8 @@ import com.apadmi.mockzilla.ui.i18n.Strings
 import com.apadmi.mockzilla.ui.ui.common.assets.DragCorner
 import com.apadmi.mockzilla.ui.ui.common.components.PlatformVerticalScrollbar
 import com.apadmi.mockzilla.ui.ui.common.theme.LocalMonoFontFamily
-import com.apadmi.mockzilla.ui.ui.common.theme.jsonHighlight
 import com.apadmi.mockzilla.ui.ui.common.theme.onSurfaceMuted
-import com.apadmi.mockzilla.ui.ui.common.widgets.endpoints.details.EndpointBodyVisualTransformation
+import com.apadmi.mockzilla.ui.ui.common.utils.formatting.BodyVisualTransformation
 
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.drop
@@ -121,7 +119,7 @@ internal fun EditorTextField(
         color = colorScheme.onSurface,
         fontFamily = monoFont,
     )
-    val outputTransformation = buildEditorOutputTransformation(mode)
+    val outputTransformation = BodyVisualTransformation.buildEditorOutputTransformation(mode)
 
     Column(
         modifier = modifier.border(
@@ -363,34 +361,3 @@ private fun EditorErrorBanner(
     }
 }
 
-@Composable
-private fun buildEditorOutputTransformation(mode: EditorMode): OutputTransformation? {
-    val colorScheme = MaterialTheme.colorScheme
-    val highlight = colorScheme.jsonHighlight
-    return remember(mode, highlight, colorScheme.onSurface, colorScheme.onSurfaceVariant) {
-        when (mode) {
-            EditorMode.Json -> EndpointBodyVisualTransformation(
-                comment = SpanStyle(color = colorScheme.onSurface.copy(alpha = 0.5f)),
-                brace = SpanStyle(color = colorScheme.onSurfaceVariant),
-                comma = SpanStyle(color = colorScheme.onSurfaceVariant),
-                colon = SpanStyle(color = colorScheme.onSurfaceVariant),
-                key = SpanStyle(color = highlight.keyColor),
-                string = SpanStyle(color = highlight.stringColor),
-                keyword = SpanStyle(color = highlight.boolColor),
-                number = SpanStyle(color = highlight.numberColor),
-                default = SpanStyle(color = colorScheme.onSurface),
-            )
-
-            EditorMode.Html -> HtmlBodyVisualTransformation(
-                bracket = SpanStyle(color = colorScheme.onSurfaceVariant),
-                tagName = SpanStyle(color = highlight.keyColor),
-                attributeName = SpanStyle(color = highlight.stringColor),
-                attributeValue = SpanStyle(color = highlight.numberColor),
-                comment = SpanStyle(color = colorScheme.onSurface.copy(alpha = 0.5f)),
-                default = SpanStyle(color = colorScheme.onSurface),
-            )
-
-            EditorMode.PlainText -> null
-        }
-    }
-}
