+++ mockzilla-management-ui/mockzilla-management-ui-common/src/commonMain/kotlin/com/apadmi/mockzilla/ui/ui/common/utils/formatting/HighlightingTransformation.kt
@@ -1,0 +1,8 @@
+package com.apadmi.mockzilla.ui.ui.common.utils.formatting
+
+import androidx.compose.foundation.text.input.OutputTransformation
+import androidx.compose.ui.text.AnnotatedString
+
+internal interface HighlightingTransformation: OutputTransformation  {
+    fun highlight(body: String): AnnotatedString
+}
