+++ mockzilla-management-ui/mockzilla-management-ui-common/src/commonMain/kotlin/com/apadmi/mockzilla/ui/ui/common/utils/formatting/JsonTokens.kt
@@ -1,4 +1,4 @@
-package com.apadmi.mockzilla.ui.ui.common.components.editor
+package com.apadmi.mockzilla.ui.ui.common.utils.formatting
 
 /**
  * @property startIndex First index of token in json
