+++ mockzilla-management-ui/mockzilla-management-ui-common/src/desktopTest/kotlin/com/apadmi/mockzilla/ui/ui/widgets/endpoints/details/JsonTokensTests.kt
@@ -1,8 +1,8 @@
 package com.apadmi.mockzilla.ui.ui.widgets.endpoints.details
 
-import com.apadmi.mockzilla.ui.ui.common.components.editor.JsonTokens
-import com.apadmi.mockzilla.ui.ui.common.components.editor.Token
-import com.apadmi.mockzilla.ui.ui.common.components.editor.TokenIndex
+import com.apadmi.mockzilla.ui.ui.common.utils.formatting.JsonTokens
+import com.apadmi.mockzilla.ui.ui.common.utils.formatting.Token
+import com.apadmi.mockzilla.ui.ui.common.utils.formatting.TokenIndex
 import kotlin.test.Test
 import kotlin.test.assertEquals
